refactor(kubernetes): colocate kubernetes infrastructure interfaces
The kubernetes infrastructure interfaces currently represent an amalgam of data received from clouddriver from each logical kind request and data from a separate request for the full resource manifest. There is also some overlap among these interfaces. Let's colocate them to improve readability of the upcoming refactor, at the top level because these logical kind interfaces are relevant beyond the details directories.
