--- app/scripts/modules/kubernetes/src/instance/details/IKubernetesInstance.ts
@@ -1,11 +1,0 @@
-import { IInstance, IMoniker } from '@spinnaker/core';
-
-export interface IKubernetesInstance extends IInstance {
-  kind: string;
-  name: string;
-  humanReadableName: string;
-  displayName: string;
-  apiVersion: string;
-  namespace: string;
-  moniker: IMoniker;
-}
