+++ app/scripts/modules/kubernetes/src/loadBalancer/index.ts
@@ -1,2 +1,1 @@
 export * from './details/details.controller';
-export * from './details/IKubernetesLoadBalancer';
