--- app/scripts/modules/kubernetes/src/serverGroup/details/IKubernetesServerGroup.ts
@@ -1,9 +1,0 @@
-import { IServerGroup } from '@spinnaker/core';
-
-export interface IKubernetesServerGroup extends IServerGroup {
-  kind: string;
-  displayName: string;
-  apiVersion: string;
-  disabled: boolean;
-  namespace: string;
-}
