--- app/scripts/modules/kubernetes/src/serverGroupManager/IKubernetesServerGroupManager.ts
@@ -1,8 +1,0 @@
-import { IServerGroupManager } from '@spinnaker/core';
-
-export interface IKubernetesServerGroupManager extends IServerGroupManager {
-  kind: string;
-  displayName: string;
-  apiVersion: string;
-  namespace: string;
-}
