+++ app/scripts/modules/kubernetes/src/serverGroupManager/index.ts
@@ -1,2 +1,1 @@
 export * from './details/details.controller';
-export * from './IKubernetesServerGroupManager';
